extract figma node conversion as a indepandant package
