+++ figma-node-conversion/converters/blend-mode.convert.ts
