+++ figma-node-conversion/converters/layout-grow.convert.ts
