+++ figma-node-conversion/converters/text-align.converter.ts
